refactor(submitters) (#3574) (patch)
# Description

refactoring of order submitters.
