+++ cg/meta/orders/api.py
@@ -11,66 +11,36 @@ import logging
 
 from cg.apps.lims import LimsAPI
 from cg.apps.osticket import OsTicket
-from cg.meta.orders.balsamic_qc_submitter import BalsamicQCSubmitter
-from cg.meta.orders.balsamic_submitter import BalsamicSubmitter
-from cg.meta.orders.balsamic_umi_submitter import BalsamicUmiSubmitter
-from cg.meta.orders.fastq_submitter import FastqSubmitter
-from cg.meta.orders.fluffy_submitter import FluffySubmitter
-from cg.meta.orders.metagenome_submitter import MetagenomeSubmitter
-from cg.meta.orders.microsalt_submitter import MicrosaltSubmitter
-from cg.meta.orders.mip_dna_submitter import MipDnaSubmitter
-from cg.meta.orders.mip_rna_submitter import MipRnaSubmitter
-from cg.meta.orders.rml_submitter import RmlSubmitter
-from cg.meta.orders.rnafusion_submitter import RnafusionSubmitter
-from cg.meta.orders.sars_cov_2_submitter import SarsCov2Submitter
-from cg.meta.orders.submitter import Submitter
 from cg.meta.orders.ticket_handler import TicketHandler
-from cg.meta.orders.tomte_submitter import TomteSubmitter
 from cg.models.orders.order import OrderIn, OrderType
+from cg.services.orders.submitters.order_submitter_registry import OrderSubmitterRegistry
 from cg.store.store import Store
 
 LOG = logging.getLogger(__name__)
 
 
-def _get_submit_handler(project: OrderType, lims: LimsAPI, status: Store) -> Submitter:
-    """Factory Method"""
-
-    submitters = {
-        OrderType.BALSAMIC: BalsamicSubmitter,
-        OrderType.BALSAMIC_QC: BalsamicQCSubmitter,
-        OrderType.BALSAMIC_UMI: BalsamicUmiSubmitter,
-        OrderType.FASTQ: FastqSubmitter,
-        OrderType.FLUFFY: FluffySubmitter,
-        OrderType.METAGENOME: MetagenomeSubmitter,
-        OrderType.MICROSALT: MicrosaltSubmitter,
-        OrderType.MIP_DNA: MipDnaSubmitter,
-        OrderType.MIP_RNA: MipRnaSubmitter,
-        OrderType.RML: RmlSubmitter,
-        OrderType.RNAFUSION: RnafusionSubmitter,
-        OrderType.SARS_COV_2: SarsCov2Submitter,
-        OrderType.TOMTE: TomteSubmitter,
-    }
-    if project in submitters:
-        return submitters[project](lims=lims, status=status)
-
-
 class OrdersAPI:
     """Orders API for accepting new samples into the system."""
 
-    def __init__(self, lims: LimsAPI, status: Store, osticket: OsTicket):
+    def __init__(
+        self,
+        lims: LimsAPI,
+        status: Store,
+        osticket: OsTicket,
+        submitter_registry: OrderSubmitterRegistry,
+    ):
         super().__init__()
         self.lims = lims
         self.status = status
         self.ticket_handler: TicketHandler = TicketHandler(osticket_api=osticket, status_db=status)
+        self.submitter_registry = submitter_registry
 
     def submit(self, project: OrderType, order_in: OrderIn, user_name: str, user_mail: str) -> dict:
         """Submit a batch of samples.
 
         Main entry point for the class towards interfaces that implements it.
         """
-        submit_handler: Submitter = _get_submit_handler(project, lims=self.lims, status=self.status)
-        submit_handler.validate_order(order=order_in)
-
+        submit_handler = self.submitter_registry.get_order_submitter(project)
         # detect manual ticket assignment
         ticket_number: str | None = TicketHandler.parse_ticket_number(order_in.name)
         if not ticket_number:
@@ -86,4 +56,4 @@ class OrdersAPI:
                 ticket_number=ticket_number,
             )
         order_in.ticket = ticket_number
-        return submit_handler.submit_order(order=order_in)
+        return submit_handler.submit_order(order_in=order_in)
