--- cg/meta/orders/balsamic_qc_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.balsamic_submitter import BalsamicSubmitter
-
-
-class BalsamicQCSubmitter(BalsamicSubmitter):
-    pass
