--- cg/meta/orders/balsamic_umi_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.balsamic_submitter import BalsamicSubmitter
-
-
-class BalsamicUmiSubmitter(BalsamicSubmitter):
-    pass
