--- cg/meta/orders/fluffy_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.pool_submitter import PoolSubmitter
-
-
-class FluffySubmitter(PoolSubmitter):
-    pass
