--- cg/meta/orders/mip_dna_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.case_submitter import CaseSubmitter
-
-
-class MipDnaSubmitter(CaseSubmitter):
-    pass
