--- cg/meta/orders/mip_rna_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.case_submitter import CaseSubmitter
-
-
-class MipRnaSubmitter(CaseSubmitter):
-    pass
