--- cg/meta/orders/rml_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.pool_submitter import PoolSubmitter
-
-
-class RmlSubmitter(PoolSubmitter):
-    pass
