+++ cg/server/ext.py
@@ -13,6 +13,10 @@ from cg.services.orders.order_service.order_service import OrderService
 from cg.services.orders.order_summary_service.order_summary_service import (
     OrderSummaryService,
 )
+from cg.services.orders.submitters.order_submitter_registry import (
+    OrderSubmitterRegistry,
+    setup_order_submitter_registry,
+)
 from cg.services.sample_service.sample_service import SampleService
 from cg.store.database import initialize_database
 from cg.store.store import Store
@@ -83,3 +87,6 @@ delivery_message_service = DeliveryMessageService(store=db, trailblazer_api=anal
 summary_service = OrderSummaryService(store=db, analysis_client=analysis_client)
 order_service = OrderService(store=db, status_service=summary_service)
 sample_service = SampleService(db)
+order_submitter_registry: OrderSubmitterRegistry = setup_order_submitter_registry(
+    lims=lims, status_db=db
+)
