+++ cg/services/orders/store_order_services/store_metagenome_order.py
@@ -1,41 +1,28 @@
-import datetime as dt
+import logging
+from datetime import datetime
 
-from cg.constants import DataDelivery
-from cg.constants.constants import Workflow
-from cg.constants.subject import Sex
+from cg.constants import Workflow, DataDelivery, Sex
 from cg.exc import OrderError
-from cg.meta.orders.lims import process_lims
-from cg.meta.orders.submitter import Submitter
 from cg.models.orders.order import OrderIn
 from cg.models.orders.sample_base import StatusEnum
-from cg.models.orders.samples import MetagenomeSample
-from cg.store.models import ApplicationVersion, Case, CaseSample, Customer, Sample
+from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
+from cg.services.orders.submitters.order_submitter import StoreOrderService
+from cg.store.models import Sample, Customer, ApplicationVersion, Case, CaseSample
+from cg.store.store import Store
+
+LOG = logging.getLogger(__name__)
 
 
-class MetagenomeSubmitter(Submitter):
-    def validate_order(self, order: OrderIn) -> None:
-        self._validate_sample_names_are_unique(samples=order.samples, customer_id=order.customer)
+class StoreMetagenomeOrderService(StoreOrderService):
+    """Storing service for metagenome orders."""
 
-    def _validate_sample_names_are_unique(
-        self, samples: list[MetagenomeSample], customer_id: str
-    ) -> None:
-        """Validate that the names of all samples are unused."""
-        customer: Customer = self.status.get_customer_by_internal_id(
-            customer_internal_id=customer_id
-        )
-        for sample in samples:
-            if sample.control:
-                continue
-            if self.status.get_sample_by_customer_and_name(
-                customer_entry_id=[customer.id], sample_name=sample.name
-            ):
-                raise OrderError(f"Sample name {sample.name} already in use")
+    def __init__(self, status_db: Store, lims_service: OrderLimsService):
+        self.status_db = status_db
+        self.lims = lims_service
 
-    def submit_order(self, order: OrderIn) -> dict:
+    def store_order(self, order: OrderIn) -> dict:
         """Submit a batch of metagenome samples."""
-        project_data, lims_map = process_lims(
-            lims_api=self.lims, lims_order=order, new_samples=order.samples
-        )
+        project_data, lims_map = self.lims.process_lims(lims_order=order, new_samples=order.samples)
         status_data = self.order_to_status(order)
         self._fill_in_sample_ids(samples=status_data["families"][0]["samples"], lims_map=lims_map)
         new_samples = self.store_items_in_status(
@@ -45,7 +32,6 @@ class MetagenomeSubmitter(Submitter):
             ticket_id=order.ticket,
             items=status_data["families"],
         )
-        self._add_missing_reads(new_samples)
         return {"project": project_data, "records": new_samples}
 
     @staticmethod
@@ -78,24 +64,24 @@ class MetagenomeSubmitter(Submitter):
         self,
         customer_id: str,
         order: str,
-        ordered: dt.datetime,
+        ordered: datetime,
         ticket_id: str,
         items: list[dict],
     ) -> list[Sample]:
         """Store samples in the status database."""
-        customer: Customer = self.status.get_customer_by_internal_id(
+        customer: Customer = self.status_db.get_customer_by_internal_id(
             customer_internal_id=customer_id
         )
         if customer is None:
             raise OrderError(f"unknown customer: {customer_id}")
         new_samples = []
-        case: Case = self.status.get_case_by_name_and_customer(
+        case: Case = self.status_db.get_case_by_name_and_customer(
             customer=customer, case_name=str(ticket_id)
         )
         case_dict: dict = items[0]
-        with self.status.session.no_autoflush:
+        with self.status_db.session.no_autoflush:
             for sample in case_dict["samples"]:
-                new_sample = self.status.add_sample(
+                new_sample = self.status_db.add_sample(
                     name=sample["name"],
                     sex=Sex.UNKNOWN,
                     comment=sample["comment"],
@@ -109,7 +95,7 @@ class MetagenomeSubmitter(Submitter):
                 new_sample.customer: Customer = customer
                 application_tag: str = sample["application"]
                 application_version: ApplicationVersion = (
-                    self.status.get_current_application_version_by_tag(tag=application_tag)
+                    self.status_db.get_current_application_version_by_tag(tag=application_tag)
                 )
                 if application_version is None:
                     raise OrderError(f"Invalid application: {sample['application']}")
@@ -117,7 +103,7 @@ class MetagenomeSubmitter(Submitter):
                 new_samples.append(new_sample)
 
                 if not case:
-                    case = self.status.add_case(
+                    case = self.status_db.add_case(
                         data_analysis=Workflow(case_dict["data_analysis"]),
                         data_delivery=DataDelivery(case_dict["data_delivery"]),
                         name=str(ticket_id),
@@ -126,14 +112,14 @@ class MetagenomeSubmitter(Submitter):
                         ticket=ticket_id,
                     )
                     case.customer = customer
-                    self.status.session.add(case)
-                    self.status.session.commit()
+                    self.status_db.session.add(case)
+                    self.status_db.session.commit()
 
-                new_relationship: CaseSample = self.status.relate_sample(
+                new_relationship: CaseSample = self.status_db.relate_sample(
                     case=case, sample=new_sample, status=StatusEnum.unknown
                 )
-                self.status.session.add(new_relationship)
+                self.status_db.session.add(new_relationship)
 
-        self.status.session.add_all(new_samples)
-        self.status.session.commit()
+        self.status_db.session.add_all(new_samples)
+        self.status_db.session.commit()
         return new_samples
