+++ tests/fixture_plugins/orders_fixtures/order_form_fixtures.py
@@ -12,6 +12,14 @@ from cg.constants.constants import FileFormat
 from cg.constants.orderforms import Orderform
 from cg.io.controller import ReadFile
 from cg.models.orders.constants import OrderType
+from cg.models.orders.order import OrderIn
+from cg.services.orders.store_order_services.store_fastq_order_service import StoreFastqOrderService
+from cg.services.orders.store_order_services.store_case_order import StoreCaseOrderService
+from cg.services.orders.store_order_services.store_metagenome_order import (
+    StoreMetagenomeOrderService,
+)
+from cg.services.orders.store_order_services.store_microbial_order import StoreMicrobialOrderService
+from cg.services.orders.store_order_services.store_pool_order import StorePoolOrderService
 
 
 def get_nr_samples_excel(orderform_path: str) -> int:
@@ -360,3 +368,119 @@ def json_order_list(
         OrderType.FLUFFY: fluffy_uploaded_json_order,
         OrderType.BALSAMIC: balsamic_uploaded_json_order,
     }
+
+
+@pytest.fixture(scope="session")
+def all_orders_to_submit(
+    balsamic_order_to_submit: dict,
+    fastq_order_to_submit: dict,
+    metagenome_order_to_submit: dict,
+    microbial_order_to_submit: dict,
+    mip_order_to_submit: dict,
+    mip_rna_order_to_submit: dict,
+    rml_order_to_submit: dict,
+    rnafusion_order_to_submit: dict,
+    sarscov2_order_to_submit: dict,
+) -> dict:
+    """Returns a dict of parsed order for each order type."""
+    return {
+        OrderType.BALSAMIC: OrderIn.parse_obj(balsamic_order_to_submit, project=OrderType.BALSAMIC),
+        OrderType.FASTQ: OrderIn.parse_obj(fastq_order_to_submit, project=OrderType.FASTQ),
+        OrderType.FLUFFY: OrderIn.parse_obj(rml_order_to_submit, project=OrderType.FLUFFY),
+        OrderType.METAGENOME: OrderIn.parse_obj(
+            metagenome_order_to_submit, project=OrderType.METAGENOME
+        ),
+        OrderType.MICROSALT: OrderIn.parse_obj(
+            microbial_order_to_submit, project=OrderType.MICROSALT
+        ),
+        OrderType.MIP_DNA: OrderIn.parse_obj(mip_order_to_submit, project=OrderType.MIP_DNA),
+        OrderType.MIP_RNA: OrderIn.parse_obj(mip_rna_order_to_submit, project=OrderType.MIP_RNA),
+        OrderType.RML: OrderIn.parse_obj(rml_order_to_submit, project=OrderType.RML),
+        OrderType.RNAFUSION: OrderIn.parse_obj(
+            rnafusion_order_to_submit, project=OrderType.RNAFUSION
+        ),
+        OrderType.SARS_COV_2: OrderIn.parse_obj(
+            sarscov2_order_to_submit, project=OrderType.SARS_COV_2
+        ),
+    }
+
+
+@pytest.fixture
+def balsamic_status_data(
+    balsamic_order_to_submit: dict, store_generic_order_service: StoreCaseOrderService
+) -> dict:
+    """Parse balsamic order example."""
+    project: OrderType = OrderType.BALSAMIC
+    order: OrderIn = OrderIn.parse_obj(obj=balsamic_order_to_submit, project=project)
+    return store_generic_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def fastq_status_data(
+    fastq_order_to_submit, store_fastq_order_service: StoreFastqOrderService
+) -> dict:
+    """Parse fastq order example."""
+    project: OrderType = OrderType.FASTQ
+    order: OrderIn = OrderIn.parse_obj(obj=fastq_order_to_submit, project=project)
+    return store_fastq_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def metagenome_status_data(
+    metagenome_order_to_submit: dict, store_metagenome_order_service: StoreMetagenomeOrderService
+) -> dict:
+    """Parse metagenome order example."""
+    project: OrderType = OrderType.METAGENOME
+    order: OrderIn = OrderIn.parse_obj(obj=metagenome_order_to_submit, project=project)
+
+    return store_metagenome_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def microbial_status_data(
+    microbial_order_to_submit: dict, store_microbial_order_service: StoreMicrobialOrderService
+) -> dict:
+    """Parse microbial order example."""
+    project: OrderType = OrderType.MICROSALT
+    order: OrderIn = OrderIn.parse_obj(obj=microbial_order_to_submit, project=project)
+    return store_microbial_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def mip_rna_status_data(
+    mip_rna_order_to_submit: dict, store_generic_order_service: StoreCaseOrderService
+) -> dict:
+    """Parse rna order example."""
+    project: OrderType = OrderType.MIP_RNA
+    order: OrderIn = OrderIn.parse_obj(obj=mip_rna_order_to_submit, project=project)
+    return store_generic_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def mip_status_data(
+    mip_order_to_submit: dict, store_generic_order_service: StoreCaseOrderService
+) -> dict:
+    """Parse scout order example."""
+    project: OrderType = OrderType.MIP_DNA
+    order: OrderIn = OrderIn.parse_obj(obj=mip_order_to_submit, project=project)
+    return store_generic_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def rml_status_data(
+    rml_order_to_submit: dict, store_pool_order_service: StorePoolOrderService
+) -> dict:
+    """Parse rml order example."""
+    project: OrderType = OrderType.RML
+    order: OrderIn = OrderIn.parse_obj(obj=rml_order_to_submit, project=project)
+    return store_pool_order_service.order_to_status(order=order)
+
+
+@pytest.fixture
+def tomte_status_data(
+    tomte_order_to_submit: dict, store_generic_order_service: StoreCaseOrderService
+) -> dict:
+    """Parse TOMTE order example."""
+    project: OrderType = OrderType.TOMTE
+    order: OrderIn = OrderIn.parse_obj(obj=tomte_order_to_submit, project=project)
+    return store_generic_order_service.order_to_status(order=order)
