+++ tests/services/orders/order_lims_service/test_order_lims_service.py
@@ -1,7 +1,7 @@
 import pytest
 
 from cg.constants import Workflow
-from cg.meta.orders.lims import build_lims_sample
+from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
 from cg.models.lims.sample import LimsSample
 from cg.models.orders.constants import OrderType
 from cg.models.orders.order import OrderIn
@@ -11,7 +11,9 @@ def test_to_lims_mip(mip_order_to_submit):
     # GIVEN a scout order for a trio
     order_data = OrderIn.parse_obj(obj=mip_order_to_submit, project=OrderType.MIP_DNA)
     # WHEN parsing the order to format for LIMS import
-    samples: list[LimsSample] = build_lims_sample(customer="cust003", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="cust003", samples=order_data.samples
+    )
 
     # THEN it should list all samples
     assert len(samples) == 4
@@ -43,7 +45,9 @@ def test_to_lims_fastq(fastq_order_to_submit):
     order_data = OrderIn.parse_obj(obj=fastq_order_to_submit, project=OrderType.FASTQ)
 
     # WHEN parsing the order to format for LIMS
-    samples: list[LimsSample] = build_lims_sample(customer="dummyCust", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="dummyCust", samples=order_data.samples
+    )
 
     # THEN should "work"
     assert len(samples) == 2
@@ -60,7 +64,9 @@ def test_to_lims_rml(rml_order_to_submit):
     order_data = OrderIn.parse_obj(obj=rml_order_to_submit, project=OrderType.RML)
 
     # WHEN parsing for LIMS
-    samples: list[LimsSample] = build_lims_sample(customer="dummyCust", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="dummyCust", samples=order_data.samples
+    )
 
     # THEN it should have found the same number of samples
     assert len(samples) == 4
@@ -78,7 +84,9 @@ def test_to_lims_microbial(microbial_order_to_submit):
     order_data = OrderIn.parse_obj(obj=microbial_order_to_submit, project=OrderType.MICROSALT)
 
     # WHEN parsing for LIMS
-    samples: list[LimsSample] = build_lims_sample(customer="cust000", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="cust000", samples=order_data.samples
+    )
     # THEN it should "work"
 
     assert len(samples) == 5
@@ -99,7 +107,9 @@ def test_to_lims_sarscov2(sarscov2_order_to_submit):
     order_data = OrderIn.parse_obj(obj=sarscov2_order_to_submit, project=OrderType.SARS_COV_2)
 
     # WHEN parsing for LIMS
-    samples: list[LimsSample] = build_lims_sample(customer="cust000", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="cust000", samples=order_data.samples
+    )
 
     # THEN it should have found the same number of samples
     assert len(samples) == 6
@@ -128,7 +138,9 @@ def test_to_lims_balsamic(balsamic_order_to_submit, project):
     order_data = OrderIn.parse_obj(obj=balsamic_order_to_submit, project=project)
 
     # WHEN parsing the order to format for LIMS import
-    samples: list[LimsSample] = build_lims_sample(customer="cust000", samples=order_data.samples)
+    samples: list[LimsSample] = OrderLimsService._build_lims_sample(
+        customer="cust000", samples=order_data.samples
+    )
     # THEN it should list all samples
 
     assert len(samples) == 1
