+++ tests/services/orders/test_validate_order_service/test_validate_generic_order.py
@@ -1,9 +1,11 @@
 import pytest
 
 from cg.exc import OrderError
-from cg.meta.orders.rnafusion_submitter import RnafusionSubmitter
 from cg.models.orders.constants import OrderType
 from cg.models.orders.order import OrderIn
+from cg.services.orders.validate_order_services.validate_case_order import (
+    ValidateCaseOrderService,
+)
 from cg.store.store import Store
 
 
@@ -15,13 +17,13 @@ def test__validate_one_sample_per_case_multiple_samples(
     ### GIVEN an RNAFUSION order where the first and last sample share the same case
     order_data = OrderIn.parse_obj(obj=rnafusion_order_to_submit, project=OrderType.RNAFUSION)
     order_data.samples[-1].family_name = order_data.samples[0].family_name
-    rnafusion_submitter = RnafusionSubmitter(status=base_store, lims=None)
+    validator = ValidateCaseOrderService(base_store)
 
     ### WHEN validating that each case has only one sample
     ### THEN an OrderError should be raised
 
     with pytest.raises(OrderError):
-        rnafusion_submitter._validate_only_one_sample_per_case(order_data.samples)
+        validator._validate_only_one_sample_per_case(order_data.samples)
 
 
 def test__validate_one_sample_per_case_unique_samples(
@@ -33,9 +35,9 @@ def test__validate_one_sample_per_case_unique_samples(
     order_data: OrderIn = OrderIn.parse_obj(
         obj=rnafusion_order_to_submit, project=OrderType.RNAFUSION
     )
-    rnafusion_submitter: RnafusionSubmitter = RnafusionSubmitter(status=base_store, lims=None)
+    validator = ValidateCaseOrderService(base_store)
 
     ### WHEN validating that each case has only one sample
-    rnafusion_submitter._validate_only_one_sample_per_case(order_data.samples)
+    validator._validate_only_one_sample_per_case(order_data.samples)
 
     ### THEN no errors should be raised
