+++ tests/services/orders/test_validate_order_service/test_validate_pool_order_service.py
@@ -3,10 +3,10 @@ import pytest
 from cg.constants import DataDelivery
 from cg.constants.constants import Workflow
 from cg.exc import OrderError
-from cg.meta.orders.pool_submitter import PoolSubmitter
 from cg.models.orders.constants import OrderType
 from cg.models.orders.order import OrderIn
 from cg.models.orders.samples import RmlSample
+from cg.services.orders.validate_order_services.validate_pool_order import ValidatePoolOrderService
 from cg.store.models import Customer
 from cg.store.store import Store
 from tests.store_helpers import StoreHelpers
@@ -17,7 +17,7 @@ def test_validate_normal_order(rml_order_to_submit: dict, base_store: Store):
     order = OrderIn.parse_obj(rml_order_to_submit, OrderType.RML)
 
     # WHEN validating the order
-    PoolSubmitter(status=base_store, lims=None).validate_order(order=order)
+    ValidatePoolOrderService(status_db=base_store).validate_order(order=order)
     # THEN it should be regarded as valid
 
 
@@ -30,7 +30,9 @@ def test_validate_case_name(rml_order_to_submit: dict, base_store: Store, helper
     for sample in order.samples:
         case = helpers.ensure_case(
             store=base_store,
-            case_name=PoolSubmitter.create_case_name(ticket=order.ticket, pool_name=sample.pool),
+            case_name=ValidatePoolOrderService.create_case_name(
+                ticket=order.ticket, pool_name=sample.pool
+            ),
             customer=customer,
             data_analysis=Workflow.FLUFFY,
             data_delivery=DataDelivery.STATINA,
@@ -41,4 +43,4 @@ def test_validate_case_name(rml_order_to_submit: dict, base_store: Store, helper
     # WHEN validating the order
     # THEN it should be regarded as invalid
     with pytest.raises(OrderError):
-        PoolSubmitter(status=base_store, lims=None).validate_order(order=order)
+        ValidatePoolOrderService(status_db=base_store).validate_order(order=order)
